ProblemSet1 reworked to pass the checks

+++ ProblemSet1/interpreter/interpreter.py
@@ -1,13 +1,3 @@
-#*
-# In a file called interpreter.py, implement a program that prompts the user for an arithmetic expression and then calculates and outputs the result as a floating-point value formatted to one decimal place. 
-# Assume that the user’s input will be formatted as x y z, with one space between x and y and one space between y and z, wherein:
-#
-#    x is an integer
-#    y is +, -, *, or /
-#    z is an integer
-#
-# *# 
-
 
 def main():
     expression = input("Expression: ")
@@ -16,16 +6,16 @@ def main():
     z = int(z)
 
     if y == "+":
-        print(x + z)
+        print(float(x + z))
 
     elif y == "-":
-        print(x - z)
+        print(float(x - z))
 
     elif y == "*":
-        print(x + z)
+        print(float(x + z))
 
     elif y == "/" and z != 0:
-        print(x / z)
+        print(float(x / z))
 
     else: 
         print("Syntax error")
